Change Con id to ulong (breaking)
Con ids should be of type gulong to avoid possible truncation of the id.

IPC ids were serialized as long ints and now unsigned longs because
they are pointers as integers. This changes fixes a bug in the library
which represented them as int types.

See [i3#2283](https://github.com/i3/i3/issues/2283) and #8 for
discussion.

+++ i3ipc-glib/i3ipc-con.h
@@ -114,7 +114,7 @@ GList *i3ipc_con_workspaces(i3ipcCon *self);
 
 i3ipcCon *i3ipc_con_find_focused(i3ipcCon *self);
 
-i3ipcCon *i3ipc_con_find_by_id(i3ipcCon *self, const gint con_id);
+i3ipcCon *i3ipc_con_find_by_id(i3ipcCon *self, const gulong con_id);
 
 i3ipcCon *i3ipc_con_find_by_window(i3ipcCon *self, const gint window_id);
